# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for vehicle evaluation service",
        "title": "Indicar API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/evaluations": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of evaluations with optional status filter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evaluations"
                ],
                "summary": "List evaluations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status (created, accepted, in_progress, completed, canceled)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Evaluation"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new vehicle evaluation request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evaluations"
                ],
                "summary": "Create evaluation",
                "parameters": [
                    {
                        "description": "Evaluation creation data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateEvaluationInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.Evaluation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/evaluations/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a specific evaluation by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evaluations"
                ],
                "summary": "Get evaluation by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Evaluation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Evaluation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update an existing evaluation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evaluations"
                ],
                "summary": "Update evaluation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Evaluation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Evaluation update data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateEvaluationInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Evaluation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/evaluations/{id}/photos": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of photos for an evaluation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evaluations"
                ],
                "summary": "List evaluation photos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Evaluation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.EvaluationPhoto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Upload a photo for an evaluation",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evaluations"
                ],
                "summary": "Upload evaluation photo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Evaluation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Photo file (max 10MB)",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.EvaluationPhoto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.Evaluation": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "evaluator_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "requester_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/entities.EvaluationStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "vehicle_make": {
                    "type": "string"
                },
                "vehicle_model": {
                    "type": "string"
                },
                "vehicle_plate": {
                    "type": "string"
                },
                "vehicle_year": {
                    "type": "integer"
                }
            }
        },
        "entities.EvaluationPhoto": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "evaluation_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "s3_bucket": {
                    "type": "string"
                },
                "s3_key": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                }
            }
        },
        "entities.EvaluationStatus": {
            "type": "string",
            "enum": [
                "created",
                "accepted",
                "in_progress",
                "completed",
                "canceled"
            ],
            "x-enum-varnames": [
                "EvaluationStatusCreated",
                "EvaluationStatusAccepted",
                "EvaluationStatusInProgress",
                "EvaluationStatusCompleted",
                "EvaluationStatusCanceled"
            ]
        },
...
        "services.CreateEvaluationInput": {
            "type": "object",
            "required": [
                "city_id",
                "vehicle_make",
                "vehicle_model"
            ],
            "properties": {
                "city_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "vehicle_make": {
                    "type": "string"
                },
                "vehicle_model": {
                    "type": "string"
                },
                "vehicle_plate": {
                    "type": "string"
                },
                "vehicle_year": {
                    "type": "integer"
                }
            }
        },
...
        "services.UpdateEvaluationInput": {
            "type": "object",
            "properties": {
                "evaluator_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}